Change ES sort parameters from tuple-list to map
The sorting keys for Elastic Search need to be supplied as a list of
objects. The previous implementation used a list of Erlang tuples,
which the Json converter library interpreted as an object! I've
changed the function that created this list so that it now returns a
list of hashmaps, which is interpreted correctly by the JSON encoder.